UnrestrictNumImpostorを二つのファイルに分割
